Add methods to create child epics
