+++ src/main/java/org/gitlab4j/api/EpicsApi.java
@@ -9,6 +9,8 @@ import javax.ws.rs.core.Form;
 import javax.ws.rs.core.GenericType;
 import javax.ws.rs.core.Response;
 
+import org.gitlab4j.api.models.ChildEpic;
+import org.gitlab4j.api.models.CreatedChildEpic;
 import org.gitlab4j.api.models.Epic;
 import org.gitlab4j.api.models.EpicIssue;
 import org.gitlab4j.api.models.EpicIssueLink;
@@ -457,5 +459,126 @@ public class EpicsApi extends AbstractApi {
         Response response = put(Response.Status.OK, form,
                 "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "issues", epicIssueId);
         return response.readEntity(new GenericType<List<EpicIssue>>() {});
+    }
+
+
+    /**
+     * Gets all child epics of an epic and the authenticated user has access to.
+     *
+     * <pre><code>GitLab Endpoint: GET /groups/:id/epics/:epic_iid/epics</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the IID of the epic to get child epics for
+     * @return a list of all child epics of the specified epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public List<ChildEpic> getChildEpics(Object groupIdOrPath, Long epicIid) throws GitLabApiException {
+        return (getChildEpics(groupIdOrPath, epicIid, getDefaultPerPage()).all());
+    }
+
+    /**
+     * Get a Pager of all child epics of an epic and the authenticated user has access to.
+     *
+     * <pre><code>GitLab Endpoint: GET /groups/:id/epics/:epic_iid/epics</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the IID of the epic to get child epics for
+     * @param itemsPerPage the number of child epics per page
+     * @return the Pager of all child epics of the specified epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public Pager<ChildEpic> getChildEpics(Object groupIdOrPath, Long epicIid, int itemsPerPage) throws GitLabApiException {
+        return (new Pager<ChildEpic>(this, ChildEpic.class, itemsPerPage, null, "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "epics"));
+    }
+
+    /**
+     * Gets all child epics of an epic and the authenticated user has access to as a Stream.
+     *
+     * <pre><code>GitLab Endpoint: GET /groups/:id/epics/:epic_iid/epics</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the IID of the epic to get child epics for
+     * @return a Stream of all child epics of the specified epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public Stream<ChildEpic> getChildEpicsStream(Object groupIdOrPath, Long epicIid) throws GitLabApiException {
+        return (getChildEpics(groupIdOrPath, epicIid, getDefaultPerPage()).stream());
+    }
+
+    /**
+     * Creates an association between two epics, designating one as the parent epic and the other as the child epic. A parent epic can have multiple child epics. If the new child epic already belonged to another epic, it is unassigned from that previous parent.
+     *
+     * <pre><code>GitLab Endpoint: POST /groups/:id/epics/:epic_iid/epics/:child_epic_id</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the Epic IID to assign the child epic to
+     * @param childEpicId the global ID of the child epic. Epic IID can’t be used because they can conflict with epics from other groups.
+     * @return an ChildEpic instance containing info on the newly assigned child epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public ChildEpic assignChildEpic(Object groupIdOrPath, Long epicIid, Long childEpicId) throws GitLabApiException {
+        Response response = post(Response.Status.CREATED, (Form)null,
+                "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "epics", childEpicId);
+        return (response.readEntity(ChildEpic.class));
+    }
+
+    /**
+     * Creates a new epic and associates it with provided parent epic.
+     *
+     * <pre><code>GitLab Endpoint: POST /groups/:id/epics/:epic_iid/epics</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the Epic IID to assign the child epic to (of the future parent epic)
+     * @param title the title of a newly created epic
+     * @param confidential whether the epic should be confidential (optional)
+     * @return an ChildEpic instance containing info on the newly created and assigned child epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public CreatedChildEpic createAndAssignChildEpic(Object groupIdOrPath, Long epicIid, String title, Boolean confidential) throws GitLabApiException {
+        Form formData = new GitLabApiForm()
+                .withParam("title", title, true)
+                .withParam("confidential", confidential);
+        Response response = post(Response.Status.CREATED, formData.asMap(),
+                "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "epics");
+        return (response.readEntity(CreatedChildEpic.class));
+    }
+
+    /**
+     * Re-order a child epic
+     *
+     * <pre><code>GitLab Endpoint: PUT /groups/:id/epics/:epic_iid/epics/:child_epic_id</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the Epic IID that the child epic is assigned to
+     * @param childEpicId the ID of the child epic. Epic IID can’t be used because they can conflict with epics from other groups.
+     * @param moveBeforeId the ID of a sibling epic that should be placed before the child epic (optional)
+     * @param moveAfterId the ID of a sibling epic that should be placed after the child epic (optional)
+     * @return a list of all child epics of the specified epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public List<ChildEpic> reOrderChildEpic(Object groupIdOrPath, Long epicIid, Long childEpicId, Long moveBeforeId, Long moveAfterId) throws GitLabApiException {
+        GitLabApiForm form = new GitLabApiForm()
+            .withParam("move_before_id", moveBeforeId)
+            .withParam("move_after_id", moveAfterId);
+        Response response = put(Response.Status.OK, form.asMap(),
+                "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "epics", childEpicId);
+        return response.readEntity(new GenericType<List<ChildEpic>>() {});
+    }
+
+    /**
+     * Unassigns a child epic from a parent epic.
+     *
+     * <pre><code>GitLab Endpoint: DELETE /groups/:id/epics/:epic_iid/epics/:child_epic_id</code></pre>
+     *
+     * @param groupIdOrPath the group ID, path of the group, or a Group instance holding the group ID or path
+     * @param epicIid the Epic IID to remove the child epic from
+     * @param childEpicId the ID of the child epic. Epic IID can’t be used because they can conflict with epics from other groups.
+     * @return an ChildEpic instance containing info on the removed child epic
+     * @throws GitLabApiException if any exception occurs
+     */
+    public ChildEpic unassignChildEpic(Object groupIdOrPath, Long epicIid, Long childEpicId) throws GitLabApiException {
+        Response response = delete(Response.Status.OK, null,
+                "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "epics", childEpicId);
+        return (response.readEntity(ChildEpic.class));
     }
 }
