+++ src/main/java/org/gitlab4j/api/models/AbstractMinimalEpic.java
@@ -1,0 +1,71 @@
+package org.gitlab4j.api.models;
+
+import org.gitlab4j.api.utils.JacksonJson;
+
+public class AbstractMinimalEpic<E extends AbstractMinimalEpic<E>> {
+
+    private Long id;
+    private Long iid;
+    private Long groupId;
+    private Long parentId;
+    private String title;
+    private String reference;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Long getIid() {
+        return iid;
+    }
+
+    public void setIid(Long iid) {
+        this.iid = iid;
+    }
+
+    public Long getGroupId() {
+        return groupId;
+    }
+
+    public void setGroupId(Long groupId) {
+        this.groupId = groupId;
+    }
+
+    public Long getParentId() {
+        return parentId;
+    }
+
+    public void setParentId(Long parentId) {
+        this.parentId = parentId;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    @SuppressWarnings("unchecked")
+    public E withTitle(String title) {
+        this.title = title;
+        return (E) (this);
+    }
+
+    public String getReference() {
+        return reference;
+    }
+
+    public void setReference(String reference) {
+        this.reference = reference;
+    }
+
+    public String toString() {
+        return (JacksonJson.toJsonString(this));
+    }
+}
