+++ src/test/resources/org/gitlab4j/api/child-epic.json
@@ -1,0 +1,39 @@
+{
+  "id": 30,
+  "iid": 5,
+  "group_id": 7,
+  "parent_id": 3,
+  "parent_iid": 8,
+  "title": "Ea cupiditate dolores ut vero consequatur quasi veniam voluptatem et non.",
+  "description": "Molestias dolorem eos vitae expedita impedit necessitatibus quo voluptatum.",
+  "state": "opened",
+  "web_url": "http://gitlab.example.com/groups/test/-/epics/5",
+  "reference": "&5",
+  "references": {
+    "short": "&5",
+    "relative": "&5",
+    "full": "test&5"
+  },
+  "author":{
+    "id": 7,
+    "name": "Pamella Huel",
+    "username": "arnita",
+    "state": "active",
+    "avatar_url": "http://www.gravatar.com/avatar/a2f5c6fcef64c9c69cb8779cb292be1b?s=80&d=identicon",
+    "web_url": "http://gitlab.example.com/arnita"
+  },
+  "start_date": "2018-07-01T00:00:00Z",
+  "due_date": "2018-07-31T00:00:00Z",
+  "created_at": "2018-07-17T13:36:22.770Z",
+  "updated_at": "2018-07-18T12:22:05.239Z",
+  "closed_at": "2018-08-18T12:22:05.239Z",
+  "labels": [],
+  "upvotes": 4,
+  "downvotes": 0,
+  "color": "#1068bf",
+  "_links":{
+    "self": "http://gitlab.example.com/api/v4/groups/7/epics/5",
+    "epic_issues": "http://gitlab.example.com/api/v4/groups/7/epics/5/issues",
+    "group":"http://gitlab.example.com/api/v4/groups/7"
+  }
+}
